% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnTree.R
\name{grnTree}
\alias{grnTree}
\title{Build GRNs for a tree of single cell clusters.}
\usage{
grnTree(tam, tnl, lambda1 = 1, lambda2 = 1)
}
\arguments{
\item{tam}{Adjacency matrix for the tree of cell clusters with tam(i,j) = 1 if and only i is the parent of j.}

\item{tnl}{A list that contains the expression of target gene (y) and TFs (X) for each cell cluster.}

\item{lambda1}{The first regularization parameter that imposed on sum_i{|b_i|_1}}

\item{lambda2}{The second regularization parameter that imposed on sum_(i,j){|b_i-b_j|_1}, i is the parent of j}
}
\value{
The solution path for the all the coefficients |b_i| with i=1,2,...,K, where K is the number of single cell clusters.
}
\description{
Build GRNs for a tree of single cell clusters.
}
\examples{
# Input:

#  tam: tree adjency matrix (0-1)^(K\timesK)
#           tam(i,j) = 1 if and only if j is a child of i
#  tnl: tree nodes list
#          node_i = list(y_i, X_i), i = 1,2,...,K

tam = matrix(0, nrow = 3, ncol = 3)
tnl = vector(mode = "list", length = nrow(tam))
lambda1 = 1.0
lambda2 = 1.0

# Specify the structure of the cell cluster tree
tam[1,2:3]=c(1,1)

# Specify the background network for each single cell cluster.
B = matrix(0.0, ncol = 3, nrow = 6)
B[1,] = 1
B[2, 2] = 0.1
B[3, 3] = 0.1
print(B)

for(ii in 1: length(tnl)){

  X = matrix (runif(120), ncol = 6)
  b = B[,ii]
  y = X\%*\%b
  tnl[[ii]] = list(y = y, X= X)
}

lambda2 = 0.0

res <- grnTree(tam, tnl, lambda1, lambda2)

plot(res, main = parse(text =  paste0('"grnTree"','~lambda[1] ==', lambda1,'*","~lambda[2] ==', lambda2)) )

lambda2 = 1.0*0.5

res <- grnTree(tam, tnl, lambda1, lambda2)

plot(res, main = parse(text =  paste0('"grnTree"','~lambda[1] ==', lambda1,'*","~lambda[2] ==', lambda2)) )


}
