% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitExp.R
\name{splitExp}
\alias{splitExp}
\title{Group the expression values of TFs and target gene according to the single cell cluster tree.}
\usage{
splitExp(y, expr_tf_matrix, cluster_groups)
}
\arguments{
\item{y}{Expression vector of the target gene in all cells.}

\item{expr_tf_matrix}{Expression matrix of the TFs in all cells with rows correspond cells and columns corresponds to TFs.}

\item{cluster_groups}{A list of cell clusters.}
}
\value{
A list of expression values for the target gene and TFs in each cell cluster.
}
\description{
Group the expression values of TFs and target gene according to the single cell cluster tree.
}
\examples{

# Specific the samples for each single cell cluster.

y  <-  runif(50)

expr_tf_matrix <- matrix(runif(50*6), ncol = 6, byrow = F)

cluster_groups <- list(sample(50, 20), sample(50, 20), sample(50, 20))

tnl <- splitExp(y, expr_tf_matrix, cluster_groups)

tam = matrix(0, nrow = 3, ncol = 3) # tree adjacency matrix with tam(i,j) = 1 if i is a parent of j.

tam[1,2:3]=c(1,1)

grns <- grnTree(tam, tnl, lambda1, lambda2)

plot(grns)

}
