% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSparse.R
\name{treeSparse}
\alias{treeSparse}
\title{Fast computation for tree-structured sparsity}
\usage{
treeSparse(tam, tnl, lambda1 = 1, lambda2 = 0)
}
\arguments{
\item{tam}{tree matrix}

\item{tnl}{tree node list}

\item{lambda1}{turning parameter alpha 1}

\item{lambda2}{turning parameter alpha 2}
}
\value{
coefficient for the tree-structured sparsity
}
\description{
Fast computation for tree-structured sparsity
}
\examples{

}
